+++ src/from-promise.ts
@@ -1,5 +1,5 @@
 import {IObservableValue, observable, action} from "mobx";
-import {IDENTITY} from "./utils";
+import {IDENTITY, deprecated} from "./utils";
 
 export type PromiseState = "pending" | "fulfilled" | "rejected";
 
@@ -19,11 +19,12 @@ class PromiseBasedObservable<T> implements IPromiseBasedObservable<T> {
         this._observable = observable(modifier(initialValue));
         promise.then(
             action("observableFromPromise-resolve", (value: T) => {
-                this._observable.set(this.modifier(value));
+                this._observable.set(value);
                 this._state.set("fulfilled");
             }),
             action("observableFromPromise-reject", (reason: any) => {
                 this._reason.set(reason);
+                this._observable.set(reason)
                 this._state.set("rejected");
             })
         );
@@ -36,6 +37,7 @@ class PromiseBasedObservable<T> implements IPromiseBasedObservable<T> {
         return this._state.get();
     }
     get reason(): any {
+        deprecated("In `fromPromise`: `.reason` is deprecated, use `.value` instead");
         return this._reason.get();
     }
 }
@@ -43,9 +45,8 @@ class PromiseBasedObservable<T> implements IPromiseBasedObservable<T> {
 /**
  * `fromPromise` takes a Promise and returns an object with 3 observable properties that track
  * the status of the promise. The returned object has the following observable properties:
- *  - `value`: either the initial value, or the value the Promise resolved to
+ *  - `value`: either the initial value, the value the Promise resolved to, or the value the Promise was rejected with. use `.state` if you need to be able to tell the difference
  *  - `state`: one of `"pending"`, `"fulfilled"` or `"rejected"`
- *  - `reason`: the reject reason if the state is `"rejected"`
  *  - `promise`: (not observable) the original promise object
  *
  * @example
@@ -55,7 +56,7 @@ class PromiseBasedObservable<T> implements IPromiseBasedObservable<T> {
  * when(
  *   () => fetchResult.state !== "pending"
  *   () => {
- *     console.log("Got ", fetchResult.reason || fetchResult.value)
+ *     console.log("Got ", fetchResult.value)
  *   }
  * )
  *
@@ -63,7 +64,7 @@ class PromiseBasedObservable<T> implements IPromiseBasedObservable<T> {
  * const myComponent = observer(({ fetchResult }) => {
  *   switch(fetchResult.state) {
  *      case "pending": return <div>Loading...</div>
- *      case "rejected": return <div>Ooops... {fetchResult.reason}</div>
+ *      case "rejected": return <div>Ooops... {fetchResult.value}</div>
  *      case "fulfilled": return <div>Gotcha: {fetchResult.value}</div>
  *   }
  * })