+++ src/utils.ts
@@ -8,3 +8,11 @@ export function invariant(cond: boolean, message = "Illegal state") {
     if (!cond)
         throw new Error("[mobx-utils] " + message);
 }
+
+const deprecatedMessages: string[] = [];
+export function deprecated(msg: string) {
+    if (deprecatedMessages.indexOf(msg) !== -1)
+        return;
+    deprecatedMessages.push(msg);
+    console.error("[mobx-utils] Deprecated: " + msg);
+}