+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 1.0.1
 
 * Improved typings of `fromPromise`, fixes #8
+* Implemented #6: the rejection reason of `fromPromise` is now stored in `.value` instead of `.reason` (which has been deprecated).